requirements.txt
pandas==2.2.2
requests==2.32.3
python-dateutil==2.9.0.post0
Jinja2==3.1.4
plotly==5.22.0 